Optimized Reading of save files
I've optimized the reading of the save files with the decompression, so
it doesn't have to decompress as much data. Someone can double check my
math at some point if they want. As long as I didn't go under the
maximum space we should be fine, but I highly doubt anyone will have
plugins with names that long.